Warning removed

+++ open9x.cpp
@@ -1341,7 +1341,7 @@ uint16_t active_functions = 0; // current max = 16 functions
 
 void evalFunctions()
 {
-  assert(sizeof(active_functions)*8 > FUNC_MAX-NUM_CHNOUT);
+  assert((int)(sizeof(active_functions)*8) > (int)(FUNC_MAX-NUM_CHNOUT));
 
   for (uint8_t i=0; i<NUM_CHNOUT; i++)
     safetyCh[i] = -128; // not defined
